chore: move utils.js to /controllers
